Drop verbosity parameter (-v)
Being verbose was intended to show progress, but as the focus is more on
packaging, verbosity is only useful for debugging. If needed, adding
[uncommitted] print() statements inside the .py scripts will do.

Removing this decreases complexity of the program, keeps the size of the
build logs down and should trivially speed up the process.

7e3f49239188e13cc87bd8203684acc35965cac7 introduced the 'pretty verbose'
which has now been dropped.