Possible alternate build steps for Snapcraft
This commit just stores the original attempt at doing things the
'proper' way by using the build system, scripts and sources.

This was unsuccessful, too slow (this project is slow at building
regardless) and was not used for the final initial snap build.

The build could complete, but there were no files for the snap. It might
had been nearly there.